Remove most remaining deepcell or tensorflow
