+++ torch_mesmer/deep_watershed.py
@@ -14,7 +14,6 @@ from skimage.morphology import remove_small_objects, h_maxima
 from skimage.morphology import disk, ball, square, cube, dilation
 from skimage.segmentation import relabel_sequential, watershed
 
-# from deepcell_toolbox.utils import erode_edges, fill_holes
 from .toolbox_utils import erode_edges, fill_holes
